Fixed classloaderror for api classe
