+++ src/api/java/com/klikli_dev/theurgy/api/TheurgyConstants.java
@@ -20,7 +20,7 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.klikli_dev.theurgy;
+package com.klikli_dev.theurgy.api;
 
 public class TheurgyConstants {
